Converted the packed Task arrays into resources
This is done to make it more convenient for me personally if I ever want to go back and fully rewrite this sonofabitch.
I'm probably going to make a tertiary program for myself that's effectively a "task creator", used for new wipes and other projects (such as a Lightkeeper Tracker). That'll be my first step into what are, effectively, mod tools. Pretty cool!